+++ roms_tools/tests/test_setup/test_grid.py
@@ -90,7 +90,6 @@ def test_plot_save_methods(tmp_path):
             str(tmp_path / file_str),
         ]:  # test for Path object and str
 
-            # Test saving without partitioning
             saved_filenames = grid.save(filepath)
             # Check if the .nc file was created
             filepath = Path(filepath).with_suffix(".nc")
@@ -99,18 +98,6 @@ def test_plot_save_methods(tmp_path):
             assert filepath.exists()
             # Clean up the .nc file
             filepath.unlink()
-
-            # Test saving with partitioning
-            saved_filenames = grid.save(filepath, np_eta=2, np_xi=5)
-
-            filepath_str = str(filepath.with_suffix(""))
-            expected_filepath_list = [
-                Path(filepath_str + f".{index}.nc") for index in range(10)
-            ]
-            assert saved_filenames == expected_filepath_list
-            for expected_filepath in expected_filepath_list:
-                assert expected_filepath.exists()
-                expected_filepath.unlink()
 
 
 def test_raise_if_domain_too_large():